+++ agent/agent.go
@@ -11,19 +11,19 @@ import (
 	"fmt"
 	"net/http"
 	"os"
+	"path/filepath"
 	"strings"
 	"sync"
 
 	time "github.com/echlebek/timeproxy"
 
 	"github.com/atlassian/gostatsd/pkg/statsd"
-	"github.com/sensu/sensu-go/agent/assetmanager"
+	"github.com/sensu/sensu-go/asset"
 	"github.com/sensu/sensu-go/command"
 	"github.com/sensu/sensu-go/handler"
 	"github.com/sensu/sensu-go/system"
 	"github.com/sensu/sensu-go/transport"
 	"github.com/sensu/sensu-go/types"
-	"github.com/sensu/sensu-go/util/path"
 	"github.com/sensu/sensu-go/util/retry"
 	"github.com/sirupsen/logrus"
 )
@@ -131,8 +131,10 @@ type SocketConfig struct {
 }
 
 // FixtureConfig provides a new Config object initialized with defaults for use
-// in tests
-func FixtureConfig() *Config {
+// in tests, as well as a cleanup function to call at the end of the test.
+func FixtureConfig() (*Config, func()) {
+	cacheDir := filepath.Join(os.TempDir(), "sensu-agent-test")
+
 	c := &Config{
 		AgentID: GetDefaultAgentID(),
 		API: &APIConfig{
@@ -140,7 +142,7 @@ func FixtureConfig() *Config {
 			Port: DefaultAPIPort,
 		},
 		BackendURLs:       []string{},
-		CacheDir:          path.SystemCacheDir("sensu-agent"),
+		CacheDir:          cacheDir,
 		Environment:       DefaultEnvironment,
 		KeepaliveInterval: DefaultKeepaliveInterval,
 		KeepaliveTimeout:  types.DefaultKeepaliveTimeout,
@@ -159,7 +161,11 @@ func FixtureConfig() *Config {
 		},
 		User: DefaultUser,
 	}
-	return c
+	return c, func() {
+		if err := os.RemoveAll(cacheDir); err != nil {
+			logger.Debugf("Error removing test agent cache dir: %s", err)
+		}
+	}
 }
 
 // NewConfig provides a new empty Config object
@@ -186,7 +192,7 @@ func GetDefaultAgentID() string {
 // An Agent receives and acts on messages from a Sensu Backend.
 type Agent struct {
 	api             *http.Server
-	assetManager    *assetmanager.Manager
+	assetManager    asset.Getter
 	backendSelector BackendSelector
 	cancel          context.CancelFunc
 	config          *Config
@@ -234,8 +240,6 @@ func NewAgent(config *Config) *Agent {
 	// We don't check for errors here and let the agent get created regardless
 	// of system info status.
 	_ = agent.refreshSystemInfo()
-	agent.assetManager = assetmanager.New(config.CacheDir, agent.getAgentEntity())
-
 	return agent
 }
 
@@ -430,15 +434,23 @@ func (a *Agent) buildTransportHeaderMap() http.Header {
 
 // Run starts the Agent.
 //
-// 1. Start a statsd server on the agent and logs the received metrics.
-// 2. Connect to the backend, return an error if unsuccessful.
-// 3. Start the socket listeners, return an error if unsuccessful.
-// 4. Start the send/receive pumps.
-// 5. Issue a keepalive immediately.
-// 6. Start refreshing system info periodically.
-// 7. Start sending periodic keepalives.
-// 8. Start the API server, shutdown the agent if doing so fails.
+// 1. Start the asset manager.
+// 2. Start a statsd server on the agent and logs the received metrics.
+// 3. Connect to the backend, return an error if unsuccessful.
+// 4. Start the socket listeners, return an error if unsuccessful.
+// 5. Start the send/receive pumps.
+// 6. Issue a keepalive immediately.
+// 7. Start refreshing system info periodically.
+// 8. Start sending periodic keepalives.
+// 9. Start the API server, shutdown the agent if doing so fails.
 func (a *Agent) Run() error {
+	var err error
+	a.entity = a.getAgentEntity()
+	a.assetManager, err = a.startAssetManager()
+	if err != nil {
+		return err
+	}
+
 	userCredentials := fmt.Sprintf("%s:%s", a.config.User, a.config.Password)
 	userCredentials = base64.StdEncoding.EncodeToString([]byte(userCredentials))
 	a.header = a.buildTransportHeaderMap()