+++ agent/check_handler_internal_test.go
@@ -26,7 +26,8 @@ func TestHandleCheck(t *testing.T) {
 		assert.FailNow("error marshaling check request")
 	}
 
-	config := FixtureConfig()
+	config, cleanup := FixtureConfig()
+	defer cleanup()
 	agent := NewAgent(config)
 	ex := &mockexecutor.MockExecutor{}
 	agent.executor = ex
@@ -55,7 +56,8 @@ func TestExecuteCheck(t *testing.T) {
 	request := &types.CheckRequest{Config: checkConfig, Issued: time.Now().Unix()}
 	checkConfig.Stdin = true
 
-	config := FixtureConfig()
+	config, cleanup := FixtureConfig()
+	defer cleanup()
 	agent := NewAgent(config)
 	ch := make(chan *transport.Message, 1)
 	agent.sendq = ch
@@ -145,7 +147,8 @@ func TestHandleTokenSubstitution(t *testing.T) {
 	request := &types.CheckRequest{Config: checkConfig, Issued: time.Now().Unix()}
 	checkConfig.Stdin = true
 
-	config := FixtureConfig()
+	config, cleanup := FixtureConfig()
+	defer cleanup()
 	config.ExtendedAttributes = []byte(`{"team":"devops"}`)
 	config.AgentID = "TestTokenSubstitution"
 	agent := NewAgent(config)
@@ -179,7 +182,8 @@ func TestHandleTokenSubstitutionNoKey(t *testing.T) {
 	request := &types.CheckRequest{Config: checkConfig, Issued: time.Now().Unix()}
 	checkConfig.Stdin = true
 
-	config := FixtureConfig()
+	config, cleanup := FixtureConfig()
+	defer cleanup()
 	config.ExtendedAttributes = []byte(`{"team":"devops"}`)
 	config.AgentID = "TestTokenSubstitution"
 	agent := NewAgent(config)
@@ -207,7 +211,8 @@ func TestHandleTokenSubstitutionNoKey(t *testing.T) {
 func TestPrepareCheck(t *testing.T) {
 	assert := assert.New(t)
 
-	config := FixtureConfig()
+	config, cleanup := FixtureConfig()
+	defer cleanup()
 	agent := NewAgent(config)
 
 	// Invalid check