+++ CHANGELOG.md
@@ -21,6 +21,7 @@ sensu/slack-handler.
 - Removed the `Error` store and type.
 
 ### Changed
+- Changed sensu-agent's internal asset manager to use BoltDB.
 - Changed sensuctl title colour to use terminal's configured default for bold
 text.
 - The backend no longer forcibly binds to localhost.